hello_word: Added tests on the lua side
Added tests to import the native lib produced by the lua side.

main.lua is the main test file, which will make all tests run

lib.lua is the file that abstracts the import of the native lib during the development process.

The tests are still weak and are intended to see if things are minimally working like a lua lib.